+++ src/ralph/models/edx/converters/xapi/__init__.py
@@ -3,4 +3,5 @@
 # flake8: noqa
 
 from .navigational import UIPageCloseToPageTerminated
+from .open_response_assessment import ORASaveSubmissionToQuestionSaved
 from .server import ServerEventToPageViewed