+++ src/ralph/models/xapi/__init__.py
@@ -3,6 +3,7 @@
 # flake8: noqa
 
 from .navigation.statements import PageTerminated, PageViewed
+from .open_response_assessment.statements import QuestionSaved
 from .video.statements import (
     VideoCompleted,
     VideoInitialized,