+++ src/ralph/models/xapi/constants.py
@@ -10,6 +10,9 @@ ACTIVITY_PAGE_DISPLAY = Literal["page"]  # pylint:disable=invalid-name
 ACTIVITY_PAGE_ID = Literal[  # pylint:disable=invalid-name
     "http://activitystrea.ms/schema/1.0/page"
 ]
+ACTIVITY_QUESTION_ID = Literal[  # pylint:disable=invalid-name
+    "http://adlnet.gov/expapi/activities/question"
+]
 
 # xAPI verbs
 VERB_COMPLETED_DISPLAY = Literal["completed"]  # pylint:disable=invalid-name
@@ -17,6 +20,7 @@ VERB_INITIALIZED_DISPLAY = Literal["initialized"]  # pylint:disable=invalid-name
 VERB_INTERACTED_DISPLAY = Literal["interacted"]  # pylint:disable=invalid-name
 VERB_PAUSED_DISPLAY = Literal["paused"]  # pylint:disable=invalid-name
 VERB_PLAYED_DISPLAY = Literal["played"]  # pylint:disable=invalid-name
+VERB_SAVED_DISPLAY = Literal["saved"]  # pylint:disable=invalid-name
 VERB_SEEKED_DISPLAY = Literal["seeked"]  # pylint:disable=invalid-name
 VERB_TERMINATED_DISPLAY = Literal["terminated"]  # pylint:disable=invalid-name
 VERB_VIEWED_DISPLAY = Literal["viewed"]  # pylint:disable=invalid-name
@@ -31,6 +35,9 @@ VERB_INITIALIZED_ID = Literal[  # pylint:disable=invalid-name
 VERB_INTERACTED_ID = Literal[  # pylint:disable=invalid-name
     "http://adlnet.gov/expapi/verbs/interacted"
 ]
+VERB_SAVED_ID = Literal[  # pylint:disable=invalid-name
+    "https://w3id.org/xapi/dod-isd/verbs/saved"
+]
 VERB_TERMINATED_ID = Literal[  # pylint:disable=invalid-name
     "http://adlnet.gov/expapi/verbs/terminated"
 ]
@@ -41,3 +48,4 @@ VERB_VIEWED_ID = Literal[  # pylint:disable=invalid-name
 EXTENSION_COURSE_ID = "http://adlnet.gov/expapi/activities/course"
 EXTENSION_MODULE_ID = "http://adlnet.gov/expapi/activities/module"
 EXTENSION_SCHOOL_ID = "https://w3id.org/xapi/acrossx/extensions/school"
+EXTENSION_RESPONSES_ID = "http://vocab.xapi.fr/extensions/responses"